+++ NICE.Identity/ClientApp/.eslintrc
@@ -1,0 +1,24 @@
+{
+	"plugins": ["prettier"],
+	"extends": [
+	  "react-app",
+	  "airbnb",
+	  "prettier"
+	],
+	"rules": {
+	  "import/no-extraneous-dependencies": 0,
+	  "react/forbid-prop-types": 0,
+	  "react/jsx-filename-extension": 0,
+	  "react/prefer-stateless-function": 1,
+	  "linebreak-style": 0,
+	  "react/sort-prop-types": [1,
+		{
+		  "callbacksLast": false,
+		  "ignoreCase": true,
+		  "requiredFirst": true,
+		  "sortShapeProp": true
+		}
+	  ],
+	  "prettier/prettier": 1
+	}
+  }
