+++ NICE.Identity/ClientApp/public/data.json
@@ -1,10 +1,110 @@
 {
-	[
-		{ make: "Toyota", model: "Celica", price: 35000
+	"data": [{
+			"make": "Toyota",
+			"model": "Celica",
+			"price": 35000
 		},
-		{ make: "Ford", model: "Mondeo", price: 32000
+		{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
 		},
-		{ make: "Porsche", model: "Boxter", price: 72000
-		}
+		{
+			"make": "Porsche",
+			"model": "Boxter",
+			"price": 72000
+		}{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},{
+			"make": "Ford",
+			"model": "Mondeo",
+			"price": 32000
+		},
 	]
 }
