+++ NICE.Identity/ClientApp/public/index.html
@@ -20,7 +20,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>WebApplication5</title>
+    <title>User Admin POC</title>
   </head>
   <body>
     <noscript>
