+++ NICE.Identity/ClientApp/src/App.jsx
@@ -5,7 +5,6 @@ import { FindUser } from './Pages/FindUser';
 
 
 export default class App extends Component {
-  displayName = App.name
 
   render() {
     return (
