--- NICE.Identity/ClientApp/src/Pages/FindUser/FindUser.js
@@ -1,21 +1,0 @@
-import React, { Component } from 'react';
-import { Col, Grid, Row } from 'react-bootstrap';
-import { FindUserControls } from '../../components/FindUserControls';
-import { GridComponent } from '../../components/Grid';
-
-export class FindUser extends Component {
-  
-
-  render() {
-    return (
-      <Grid fluid>
-        <Row>
-           <FindUserControls />
-        </Row>
-		<Row>
-			<GridComponent />
-		</Row>
-      </Grid>
-    );
-  }
-}
