+++ NICE.Identity/ClientApp/src/Pages/FindUser/components.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import { Row } from 'react-bootstrap';
+
+export const Container = styled(Row)`
+	min-height: 12.5rem;
+	border: 1px solid #666;
+	margin: 5rem 0 5rem 0;
+`
