+++ NICE.Identity/ClientApp/src/components/FindUserControls/index.jsx
