+++ NICE.Identity/ClientApp/src/components/Grid/components.js
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 
-export const GreedBorderDiv = styled.div`
-	border: 1px solid green;
+export const GridComponent = styled.div `
+	width: 100%;
+	height:50rem;
 `;
