+++ NICE.Identity/ClientApp/src/components/Layout/Layout.jsx
