+++ NICE.Identity/ClientApp/src/index.css
@@ -1,6 +1,10 @@
 @media (max-width: 767px) {
     /* On small screens, the nav menu spans the full width of the screen. Leave a space for it. */
     body {
+        font-size: 16px;
         padding-top: 50px;
+    }
+    a{
+        text-decoration: none;
     }
 }
